推荐系统完善/定时表征生成并更新rank

+++ provider/src/main/java/cn/edu/neu/shop/pin/mapper/PinUserProductCommentMapper.java
@@ -21,5 +21,5 @@ public interface PinUserProductCommentMapper extends BaseMapper<PinUserProductCo
 
     void updateMerchantComment(Integer commentId, String commentContent, Date commentTime);
 
-    Integer getAvgScore(Integer id);
+    Double getAvgScore(Integer id);
 }
